backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
feedparser==6.0.11
anthropic==0.23.1
pydantic==2.6.0
sqlalchemy==2.0.25
aiosqlite==0.20.0
python-dotenv==1.0.0
lxml==5.1.0
